requirements.txt
Flask
Flask-Session
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
authlib
mysql-connector-python
qrcode
python-dotenv
Pillow

